Bunch of stuff (#70)
* Update hierarchy of folders in db
Add start and end for lp tasks and filter on it in the big query computing points
Create scripts to add new tasks for lp_points
Create new script to truncate only tables that we want
Add creation date & block on usg_markets
Update start => start_date and closed=> closed_date
Fix error in Onchain Tx bot that was canceling the notification to Telegram

* Add create_date and block for markets creation
Fix typo in renaming _date

* Fix linter

* Update reset_lp_points

* Fix truncate script

* Fix truncate

* Fix truncate

* Add logs

* Fix indexer block

* Fix boost seeding

* Update query with bigint to dont overflow
add console.log to debug date

* All the points in bigint in hell query

* Fix more int => bigint

* Fix add gauge script

---------

+++ src/sql-functions/get_user_points_per_task.sql
@@ -6,20 +6,20 @@ CREATE OR REPLACE FUNCTION points.get_user_points_per_task(
 RETURNS TABLE (
   user_address    varchar,
   task_id         bigint,
-  points          int,
-  booster_points  int,
+  points          bigint,
+  booster_points  bigint,
   godfather_id    bigint,
-  godfather_points int
+  godfather_points bigint
 )
 LANGUAGE sql
 AS $$
   SELECT
     d.user_address,
     d.task_id,
-    SUM(d.base_points)::int     AS points,
-    SUM(d.booster_points)::int  AS booster_points,
+    SUM(d.base_points)::bigint     AS points,
+    SUM(d.booster_points)::bigint  AS booster_points,
     d.godfather_id,
-    SUM(d.godfather_points)::int AS godfather_points
+    SUM(d.godfather_points)::bigint AS godfather_points
   FROM points.get_user_points_details(start_at, end_at) d
   GROUP BY d.user_address, d.task_id, d.godfather_id;
 $$;
